Financial_Transactions_Fraud_Detection/requirements.txt
pandas==2.1.0
numpy==1.26.2
matplotlib==3.8.1
seaborn==0.12.3
scikit-learn==1.3.2
joblib==1.3.2
python-dotenv==1.0.1
mysql-connector-python==8.1.0
